add enpooint and update page pembayaran

+++ laravel/larashop-apis/database/migrations/2020_05_26_030538_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('user_id');
-            $table->float('total_price');
+            $table->double('total_price');
             $table->string('invoice_number');
             $table->string('courier_service')->nullable();
             $table->enum('status', ['SUBMIT', 'PROCESS', 'FINISH',
